Move methods update check methods from GameRunner to 'UpdateChecks'
