+++ game-core/src/main/java/games/strategy/engine/framework/EngineVersionProperties.java
@@ -23,7 +23,7 @@ import games.strategy.net.OpenFileUtility;
 import games.strategy.triplea.UrlConstants;
 import games.strategy.util.Version;
 
-class EngineVersionProperties {
+public class EngineVersionProperties {
   private static final String TRIPLEA_VERSION_LINK =
       "https://raw.githubusercontent.com/triplea-game/triplea/master/latest_version.properties";
   private final Version latestVersionOut;
@@ -43,7 +43,7 @@ class EngineVersionProperties {
     changelogLink = props.getProperty("CHANGELOG", UrlConstants.RELEASE_NOTES.toString());
   }
 
-  static EngineVersionProperties contactServerForEngineVersionProperties() {
+  public static EngineVersionProperties contactServerForEngineVersionProperties() {
     // sourceforge sometimes takes a long while to return results
     // so run a couple requests in parallel, starting with delays to try and get a response quickly
     final AtomicReference<EngineVersionProperties> ref = new AtomicReference<>();
@@ -81,7 +81,7 @@ class EngineVersionProperties {
     return props;
   }
 
-  Version getLatestVersionOut() {
+  public Version getLatestVersionOut() {
     return latestVersionOut;
   }
 
@@ -119,7 +119,7 @@ class EngineVersionProperties {
             + "</body></html>";
   }
 
-  Component getOutOfDateComponent() {
+  public Component getOutOfDateComponent() {
     final JPanel panel = new JPanel(new BorderLayout());
     final JEditorPane intro = new JEditorPane("text/html", getOutOfDateMessage());
     intro.setEditable(false);
